Added items, finished harvesting
Added in the gold, stone and wood items properly. They have an ID that matches the resource they are harvested from, allowing the harvesting to be tracked. Harvesting spams the data to the screen currently, but can be easily modified to show nothing until an inventory UI is built out.