Finalize World Interactions
System is complete to prevent players from performing certain actions based on held item slot. Added a null cursor for when a player hovers over a non-interactable, which is set to switch when the appropriate tool is selected.
Since the only interactables are currently resources, hovering over them with anything but the pickaxe will produce a red X cursor, and will prevent collecting resources until the pickaxe is held.